+++ src/test/java/org/jabref/logic/citationkeypattern/BracketedPatternTest.java
@@ -608,14 +608,44 @@ class BracketedPatternTest {
     @Test
     void expandBracketsWithModifierContainingRegexCharacterClass() {
         BibEntry bibEntry = new BibEntry().withField(StandardField.TITLE, "Wickedness:Managing");
-
         assertEquals("Wickedness.Managing", BracketedPattern.expandBrackets("[title:regex(\"[:]+\",\".\")]", null, bibEntry, null));
+    }
+
+    @Test
+    void regExForFirstWord() {
+        BibEntry bibEntry = new BibEntry().withField(StandardField.TITLE, "First Second Third");
+        assertEquals("First", BracketedPattern.expandBrackets("[TITLE:regex(\"(\\w+).*\",\"$1\")]", null, bibEntry, null));
+    }
+
+    @Test
+    void regExWithComma() {
+        BibEntry bibEntry = new BibEntry().withField(StandardField.TITLE, "First,Second,Third");
+        assertEquals("First+Second+Third", BracketedPattern.expandBrackets("[TITLE:regex(\",\",\"+\")]", null, bibEntry, null));
+    }
+
+    @Test
+    void regExWithEscapedQuote() {
+        BibEntry bibEntry = new BibEntry().withField(StandardField.TITLE, "First\"Second\"Third");
+        assertEquals("First+Second+Third", BracketedPattern.expandBrackets("[TITLE:regex(\"\\\"\",\"+\")]", null, bibEntry, null));
+    }
+
+    @Test
+    void regExWithEtAlTwoAuthors() {
+        // Example from https://docs.jabref.org/setup/citationkeypatterns#modifiers
+        BibEntry bibEntry = new BibEntry().withField(StandardField.AUTHOR, "First Last and Second Last");
+        assertEquals("LastAndLast", BracketedPattern.expandBrackets("[auth.etal:regex(\"\\.etal\",\"EtAl\"):regex(\"\\.\",\"And\")]", null, bibEntry, null));
+    }
+
+    @Test
+    void regExWithEtAlThreeAuthors() {
+        // Example from https://docs.jabref.org/setup/citationkeypatterns#modifiers
+        BibEntry bibEntry = new BibEntry().withField(StandardField.AUTHOR, "First Last and Second Last and Third Last");
+        assertEquals("LastEtAl", BracketedPattern.expandBrackets("[auth.etal:regex(\"\\.etal\",\"EtAl\"):regex(\"\\.\",\"And\")]", null, bibEntry, null));
     }
 
     @Test
     void expandBracketsEmptyStringFromEmptyBrackets() {
         BibEntry bibEntry = new BibEntry();
-
         assertEquals("", BracketedPattern.expandBrackets("[]", null, bibEntry, null));
     }
 
@@ -623,7 +653,6 @@ class BracketedPatternTest {
     void expandBracketsInstitutionAbbreviationFromProvidedAbbreviation() {
         BibEntry bibEntry = new BibEntry()
                 .withField(StandardField.AUTHOR, "{European Union Aviation Safety Agency ({EUASABRACKET})}");
-
         assertEquals("EUASABRACKET", BracketedPattern.expandBrackets("[auth]", null, bibEntry, null));
     }
 
@@ -631,7 +660,6 @@ class BracketedPatternTest {
     void expandBracketsInstitutionAbbreviationForAuthorContainingUnion() {
         BibEntry bibEntry = new BibEntry()
                 .withField(StandardField.AUTHOR, "{European Union Aviation Safety Agency}");
-
         assertEquals("EUASA", BracketedPattern.expandBrackets("[auth]", null, bibEntry, null));
     }
 
@@ -639,7 +667,6 @@ class BracketedPatternTest {
     void expandBracketsLastNameForAuthorStartingWithOnlyLastNameStartingWithLowerCase() {
         BibEntry bibEntry = new BibEntry()
                 .withField(StandardField.AUTHOR, "{eBay}");
-
         assertEquals("eBay", BracketedPattern.expandBrackets("[auth]", null, bibEntry, null));
     }
 
@@ -647,7 +674,6 @@ class BracketedPatternTest {
     void expandBracketsLastNameWithChineseCharacters() {
         BibEntry bibEntry = new BibEntry()
                 .withField(StandardField.AUTHOR, "杨秀群");
-
         assertEquals("杨秀群", BracketedPattern.expandBrackets("[auth]", null, bibEntry, null));
     }
 
@@ -655,7 +681,6 @@ class BracketedPatternTest {
     void expandBracketsUnmodifiedStringFromLongFirstPageNumber() {
         BibEntry bibEntry = new BibEntry()
                 .withField(StandardField.PAGES, "2325967120921344");
-
         assertEquals("2325967120921344", BracketedPattern.expandBrackets("[firstpage]", null, bibEntry, null));
     }
 
@@ -663,7 +688,6 @@ class BracketedPatternTest {
     void expandBracketsUnmodifiedStringFromLongLastPageNumber() {
         BibEntry bibEntry = new BibEntry()
                 .withField(StandardField.PAGES, "2325967120921344");
-
         assertEquals("2325967120921344", BracketedPattern.expandBrackets("[lastpage]", null, bibEntry, null));
     }